Adrienne's moment mapping procedures.  From the HI expert herself:
If it's the wrong colorbar.pro let me know, but I'm pretty sure it's the
correct one.

You'll need read_cube_fits to read in the file. Then run:

moments, <cube>, <cube header>, /doplot

and that should do it. you might want to set nl0 = 20 to get more
levels on the moment maps (or nl1, nl2, etc.) where that's <number of
levels for moment zero> = 20 (or <number of levels for moment 1> =
<some number>).